move BatchEvaluationScanner to types package and make type constructor package private, and use scanner in tests
